+++ include/ui/setup.hpp
@@ -6,16 +6,17 @@
 namespace nwo5::ui {
     template<typename ImplNode, typename Node = std::remove_pointer_t<ImplNode>>
     requires std::derived_from<Node, cocos2d::CCNode>
-    class Create final {
+    class Setup final {
     private:
         Node* m_node;
 
     public:
         using Type = Node;
 
-        template<typename T>
-        requires std::derived_from<T, Node>
-        operator Node* const() const {
+        Setup(Node* node) 
+            : m_node(node) {}
+
+        operator Node*() const {
             return m_node;
         }
 
@@ -26,12 +27,12 @@ namespace nwo5::ui {
             return m_node;
         }
 
-        Create parent(cocos2d::CCNode* pNode) requires std::derived_from<Node, cocos2d::CCNode> {
+        Setup parent(cocos2d::CCNode* pNode) requires std::derived_from<Node, cocos2d::CCNode> {
             pNode->addChild(m_node);
             return {m_node};
         }
         template<typename... Args>
-        Create children(Args... pChildren) requires std::derived_from<Node, cocos2d::CCNode> {
+        Setup children(Args... pChildren) requires std::derived_from<Node, cocos2d::CCNode> {
             (m_node->addChild(pChildren), ...);
 
             if constexpr (std::derived_from<Node, cocos2d::CCMenu>) {
@@ -41,62 +42,62 @@ namespace nwo5::ui {
             return {m_node};
         }
 
-        Create pos(cocos2d::CCPoint pPos) requires std::derived_from<Node, cocos2d::CCNode> {
+        Setup pos(cocos2d::CCPoint pPos) requires std::derived_from<Node, cocos2d::CCNode> {
             m_node->setPosition(pPos);
             return {m_node};
         }
-        Create pos(float pX, float pY) requires std::derived_from<Node, cocos2d::CCNode> {
+        Setup pos(float pX, float pY) requires std::derived_from<Node, cocos2d::CCNode> {
             return pos({pX, pY});
         }
-        Create pos(cocos2d::CCNode* pCopy) requires std::derived_from<Node, cocos2d::CCNode> {
+        Setup pos(cocos2d::CCNode* pCopy) requires std::derived_from<Node, cocos2d::CCNode> {
             return pos(pCopy->getPosition());
         }
-        Create posX(float pX) requires std::derived_from<Node, cocos2d::CCNode> {
+        Setup posX(float pX) requires std::derived_from<Node, cocos2d::CCNode> {
             m_node->setPositionX(pX);
             return {m_node};
         }
-        Create posX(cocos2d::CCNode* pCopy) requires std::derived_from<Node, cocos2d::CCNode> {
+        Setup posX(cocos2d::CCNode* pCopy) requires std::derived_from<Node, cocos2d::CCNode> {
             return posX(pCopy->getPositionX());
         }
-        Create posY(float pY) requires std::derived_from<Node, cocos2d::CCNode> {
+        Setup posY(float pY) requires std::derived_from<Node, cocos2d::CCNode> {
             m_node->setPositionY(pY);
             return {m_node};
         }
-        Create posY(cocos2d::CCNode* pCopy) requires std::derived_from<Node, cocos2d::CCNode> {
+        Setup posY(cocos2d::CCNode* pCopy) requires std::derived_from<Node, cocos2d::CCNode> {
             return posY(pCopy->getPositionY());
         }
 
-        Create move(cocos2d::CCPoint pOffset) requires std::derived_from<Node, cocos2d::CCNode> {
+        Setup move(cocos2d::CCPoint pOffset) requires std::derived_from<Node, cocos2d::CCNode> {
             return pos(m_node->getPosition() + pOffset);
         }
-        Create move(float pX, float pY) requires std::derived_from<Node, cocos2d::CCNode> {
+        Setup move(float pX, float pY) requires std::derived_from<Node, cocos2d::CCNode> {
             return move({pX, pY});
         }
-        Create moveX(float pX) requires std::derived_from<Node, cocos2d::CCNode> {
+        Setup moveX(float pX) requires std::derived_from<Node, cocos2d::CCNode> {
             return posX(m_node->getPositionX() + pX);
         }
-        Create moveY(float pY) requires std::derived_from<Node, cocos2d::CCNode> {
+        Setup moveY(float pY) requires std::derived_from<Node, cocos2d::CCNode> {
             return posY(m_node->getPositionY() + pY);
         }
 
-        Create above(cocos2d::CCNode* pNode, float pGap = 0.0f, DirectionAlignment pAlignment = DirectionAlignment::Center) requires std::derived_from<Node, cocos2d::CCNode> {
+        Setup above(cocos2d::CCNode* pNode, float pGap = 0.0f, DirectionAlignment pAlignment = DirectionAlignment::Center) requires std::derived_from<Node, cocos2d::CCNode> {
             nwo5::ui::positionAbove(m_node, pNode, pGap, pAlignment);
             return {m_node};
         }
-        Create below(cocos2d::CCNode* pNode, float pGap = 0.0f, DirectionAlignment pAlignment = DirectionAlignment::Center) requires std::derived_from<Node, cocos2d::CCNode> {
+        Setup below(cocos2d::CCNode* pNode, float pGap = 0.0f, DirectionAlignment pAlignment = DirectionAlignment::Center) requires std::derived_from<Node, cocos2d::CCNode> {
             nwo5::ui::positionBelow(m_node, pNode, pGap, pAlignment);
             return {m_node};
         }
-        Create right(cocos2d::CCNode* pNode, float pGap = 0.0f, DirectionAlignment pAlignment = DirectionAlignment::Center) requires std::derived_from<Node, cocos2d::CCNode> {
+        Setup right(cocos2d::CCNode* pNode, float pGap = 0.0f, DirectionAlignment pAlignment = DirectionAlignment::Center) requires std::derived_from<Node, cocos2d::CCNode> {
             nwo5::ui::positionRight(m_node, pNode, pGap, pAlignment);
             return {m_node};
         }
-        Create left(cocos2d::CCNode* pNode, float pGap = 0.0f, DirectionAlignment pAlignment = DirectionAlignment::Center) requires std::derived_from<Node, cocos2d::CCNode> {
+        Setup left(cocos2d::CCNode* pNode, float pGap = 0.0f, DirectionAlignment pAlignment = DirectionAlignment::Center) requires std::derived_from<Node, cocos2d::CCNode> {
             nwo5::ui::positionLeft(m_node, pNode, pGap, pAlignment);
             return {m_node};
         }
 
-        Create center() {
+        Setup center() {
             if (m_node->getParent()) {
                 return pos(m_node->getParent()->getContentSize() / 2);
             }
@@ -105,26 +106,26 @@ namespace nwo5::ui {
             }
         }
 
-        Create anchor(cocos2d::CCPoint pAnchor) requires std::derived_from<Node, cocos2d::CCNode> {
+        Setup anchor(cocos2d::CCPoint pAnchor) requires std::derived_from<Node, cocos2d::CCNode> {
             m_node->setAnchorPoint(pAnchor);
             return {m_node};
         }
-        Create anchor(float pX, float pY) requires std::derived_from<Node, cocos2d::CCNode> {
+        Setup anchor(float pX, float pY) requires std::derived_from<Node, cocos2d::CCNode> {
             return anchor({pX, pY});
         }
-         Create anchor(cocos2d::CCNode* pCopy) requires std::derived_from<Node, cocos2d::CCNode> {
+         Setup anchor(cocos2d::CCNode* pCopy) requires std::derived_from<Node, cocos2d::CCNode> {
             return anchor(pCopy->getAnchorPoint());
         }
 
-        Create rotation(float pRotation) requires std::derived_from<Node, cocos2d::CCNode> {
+        Setup rotation(float pRotation) requires std::derived_from<Node, cocos2d::CCNode> {
             m_node->setRotation(pRotation);
             return {m_node};
         }
-        Create rotation(cocos2d::CCNode* pCopy) requires std::derived_from<Node, cocos2d::CCNode> {
+        Setup rotation(cocos2d::CCNode* pCopy) requires std::derived_from<Node, cocos2d::CCNode> {
             return rotation(pCopy->getRotation());
         }
 
-        Create scale(float pScale) requires std::derived_from<Node, cocos2d::CCNode> {
+        Setup scale(float pScale) requires std::derived_from<Node, cocos2d::CCNode> {
             if constexpr (std::derived_from<Node, CCMenuItemSpriteExtra>) {
                 m_node->getNormalImage()->setScale(pScale);
                 m_node->updateSprite();
@@ -134,10 +135,10 @@ namespace nwo5::ui {
             }
             return {m_node};
         }
-        Create scale(cocos2d::CCNode* pCopy) requires std::derived_from<Node, cocos2d::CCNode> {
+        Setup scale(cocos2d::CCNode* pCopy) requires std::derived_from<Node, cocos2d::CCNode> {
             return scale(pCopy->getScale());
         }
-        Create scale(float pX, float pY) requires std::derived_from<Node, cocos2d::CCNode> {
+        Setup scale(float pX, float pY) requires std::derived_from<Node, cocos2d::CCNode> {
             if constexpr (std::derived_from<Node, CCMenuItemSpriteExtra>) {
                 auto image = m_node->getNormalImage();
                 image->setScaleX(pX);
@@ -150,7 +151,7 @@ namespace nwo5::ui {
             }
             return {m_node};
         }
-        Create scaleX(float pX) requires std::derived_from<Node, cocos2d::CCNode> {
+        Setup scaleX(float pX) requires std::derived_from<Node, cocos2d::CCNode> {
             if constexpr (std::derived_from<Node, CCMenuItemSpriteExtra>) {
                 m_node->getNormalImage()->setScaleX(pX);
                 m_node->updateSprite();
@@ -160,10 +161,10 @@ namespace nwo5::ui {
             }
             return {m_node};
         }
-        Create scaleX(cocos2d::CCNode* pCopy) requires std::derived_from<Node, cocos2d::CCNode> {
+        Setup scaleX(cocos2d::CCNode* pCopy) requires std::derived_from<Node, cocos2d::CCNode> {
             return scaleX(pCopy->getScaleX());
         }
-        Create scaleY(float pY) requires std::derived_from<Node, cocos2d::CCNode> {
+        Setup scaleY(float pY) requires std::derived_from<Node, cocos2d::CCNode> {
             if constexpr (std::derived_from<Node, CCMenuItemSpriteExtra>) {
                 m_node->getNormalImage()->setScaleY(pY);
                 m_node->updateSprite();
@@ -173,11 +174,11 @@ namespace nwo5::ui {
             }
             return {m_node};
         }
-        Create scaleY(cocos2d::CCNode* pCopy) requires std::derived_from<Node, cocos2d::CCNode> {
+        Setup scaleY(cocos2d::CCNode* pCopy) requires std::derived_from<Node, cocos2d::CCNode> {
             return scaleY(pCopy->getScaleY());
         }
 
-        Create scaleToFit(float pSize) requires std::derived_from<Node, cocos2d::CCNode> {
+        Setup scaleToFit(float pSize) requires std::derived_from<Node, cocos2d::CCNode> {
             if (const auto origSize = std::max(m_node->getContentWidth(), m_node->getContentHeight())) {
                 return scale(pSize / origSize);
             }
@@ -185,7 +186,7 @@ namespace nwo5::ui {
                 return {m_node};
             }
         }
-        Create scaleWidthToFit(float pSize) requires std::derived_from<Node, cocos2d::CCNode> {
+        Setup scaleWidthToFit(float pSize) requires std::derived_from<Node, cocos2d::CCNode> {
             const auto nodeSize = m_node->getContentSize();
 
             if (const auto origSize = m_node->getContentWidth()) {
@@ -195,7 +196,7 @@ namespace nwo5::ui {
                 return {m_node};
             }
         }
-        Create scaleHeightToFit(float pSize) requires std::derived_from<Node, cocos2d::CCNode> {
+        Setup scaleHeightToFit(float pSize) requires std::derived_from<Node, cocos2d::CCNode> {
             const auto nodeSize = m_node->getContentSize();
 
             if (const auto origSize = m_node->getContentHeight()) {
@@ -205,7 +206,7 @@ namespace nwo5::ui {
                 return {m_node};
             }
         }
-        Create limitScaleToFit(float pSize) requires std::derived_from<Node, cocos2d::CCNode> {
+        Setup limitScaleToFit(float pSize) requires std::derived_from<Node, cocos2d::CCNode> {
             const auto origSize = std::max(m_node->getContentWidth(), m_node->getContentHeight());
             const auto origScale = std::max(m_node->getScaleX(), m_node->getScaleY());
 
@@ -216,7 +217,7 @@ namespace nwo5::ui {
                 return {m_node};
             }
         }
-        Create limitScaleWidthToFit(float pSize) requires std::derived_from<Node, cocos2d::CCNode> {
+        Setup limitScaleWidthToFit(float pSize) requires std::derived_from<Node, cocos2d::CCNode> {
             const auto origSize = m_node->getContentWidth();
                 const auto origScale = m_node->getScaleX();
 
@@ -227,7 +228,7 @@ namespace nwo5::ui {
                 return {m_node};
             }
         }
-        Create limitScaleHeightToFit(float pSize) requires std::derived_from<Node, cocos2d::CCNode> {
+        Setup limitScaleHeightToFit(float pSize) requires std::derived_from<Node, cocos2d::CCNode> {
             const auto origSize = m_node->getContentHeight();
             const auto origScale = m_node->getScaleY();
 
@@ -239,152 +240,172 @@ namespace nwo5::ui {
             }
         }
 
-        Create size(cocos2d::CCSize pSize) requires std::derived_from<Node, cocos2d::CCNode> {
+        Setup size(cocos2d::CCSize pSize) requires std::derived_from<Node, cocos2d::CCNode> {
             m_node->setContentSize(pSize);
             return {m_node};
         }
-        Create size(float pWidth, float pHeight) requires std::derived_from<Node, cocos2d::CCNode> {
+        Setup size(float pWidth, float pHeight) requires std::derived_from<Node, cocos2d::CCNode> {
             m_node->setContentSize({pWidth, pHeight});
             return {m_node};
         }
-        Create size(cocos2d::CCNode* pCopy) requires std::derived_from<Node, cocos2d::CCNode> {
+        Setup size(cocos2d::CCNode* pCopy) requires std::derived_from<Node, cocos2d::CCNode> {
             return size(pCopy->getContentSize());
         }
 
-        Create width(float pWidth) requires std::derived_from<Node, cocos2d::CCNode> {
+        Setup width(float pWidth) requires std::derived_from<Node, cocos2d::CCNode> {
             m_node->setContentWidth(pWidth);
             return {m_node};
         }
-        Create width(cocos2d::CCNode* pCopy) requires std::derived_from<Node, cocos2d::CCNode> {
+        Setup width(cocos2d::CCNode* pCopy) requires std::derived_from<Node, cocos2d::CCNode> {
             return width(pCopy->getContentWidth());
         }
-        Create height(float pHeight) requires std::derived_from<Node, cocos2d::CCNode> {
+        Setup height(float pHeight) requires std::derived_from<Node, cocos2d::CCNode> {
             m_node->setContentHeight(pHeight);
             return {m_node};
         }
-        Create height(cocos2d::CCNode* pCopy) requires std::derived_from<Node, cocos2d::CCNode> {
+        Setup height(cocos2d::CCNode* pCopy) requires std::derived_from<Node, cocos2d::CCNode> {
             return height(pCopy->getContentHeight());
         }
 
-        Create tag(int pTag) requires std::derived_from<Node, cocos2d::CCNode> {
+        Setup tag(int pTag) requires std::derived_from<Node, cocos2d::CCNode> {
             m_node->setTag(pTag);
             return {m_node};
         }
 
-        Create id(std::string_view pID) requires std::derived_from<Node, cocos2d::CCNode> {
+        Setup id(const std::string& pID) requires std::derived_from<Node, cocos2d::CCNode> {
             m_node->setID(pID);
             return {m_node};
         }
         template <typename T, typename... Args>
-        Create id(fmt::format_string<T, Args...> pFormat, T&& pArg, Args&&... pArgs) requires std::derived_from<T, cocos2d::CCNode> {
+        Setup id(fmt::format_string<T, Args...> pFormat, T&& pArg, Args&&... pArgs) requires std::derived_from<T, cocos2d::CCNode> {
             m_node->setID(fmt::format(pFormat, std::forward<T>(pArg), std::forward<Args>(pArgs)...));
             return {m_node};
         }
 
-        Create order(int pOrder) requires std::derived_from<Node, cocos2d::CCNode> {
+        Setup order(int pOrder) requires std::derived_from<Node, cocos2d::CCNode> {
             m_node->setZOrder(pOrder);
             return {m_node};
         }
-        Create order(cocos2d::CCNode* pCopy) requires std::derived_from<Node, cocos2d::CCNode> {
+        Setup order(cocos2d::CCNode* pCopy) requires std::derived_from<Node, cocos2d::CCNode> {
             m_node->setZOrder(pCopy->getZOrder());
             return {m_node};
         }
 
-        Create visible(bool pVisible) requires std::derived_from<Node, cocos2d::CCNode> {
+        Setup visible(bool pVisible) requires std::derived_from<Node, cocos2d::CCNode> {
             m_node->setVisible(pVisible);
             return {m_node};
         }
-        Create visible(cocos2d::CCNode* pCopy) requires std::derived_from<Node, cocos2d::CCNode> {
+        Setup visible(cocos2d::CCNode* pCopy) requires std::derived_from<Node, cocos2d::CCNode> {
             return visible(pCopy->isVisible());
         }
-        Create hide() requires std::derived_from<Node, cocos2d::CCNode> {
+        Setup hide() requires std::derived_from<Node, cocos2d::CCNode> {
             m_node->setVisible(false);
             return {m_node};
         }
 
 
 
-        Create color(cocos2d::ccColor3B const& pColor) requires std::derived_from<Node, cocos2d::CCNodeRGBA> {
+        Setup flipX(bool pFlip = true) requires std::derived_from<Node, cocos2d::CCNode> {
+            if constexpr (std::derived_from<Node, cocos2d::CCSprite>) {
+                m_node->setFlipX(pFlip);
+            }
+            else {
+                m_node->setRotationY(pFlip ? 180.0f : 0.0f);
+            }
+            return {m_node};
+        }
+        Setup flipY(bool pFlip = true) requires std::derived_from<Node, cocos2d::CCNode> {
+            if constexpr (std::derived_from<Node, cocos2d::CCSprite>) {
+                m_node->setFlipY(pFlip);
+            }
+            else {
+                m_node->setRotationX(pFlip ? 180.0f : 0.0f);
+            }
+            return {m_node};
+        }
+
+
+
+        Setup color(cocos2d::ccColor3B const& pColor) requires std::derived_from<Node, cocos2d::CCRGBAProtocol> {
             m_node->setColor(pColor);
             return {m_node};
         }
-        Create color(cocos2d::CCNodeRGBA* pCopy) requires std::derived_from<Node, cocos2d::CCNodeRGBA> {
+        Setup color(Node* pCopy) requires std::derived_from<Node, cocos2d::CCRGBAProtocol> {
             return color(pCopy->getColor());
         }
-        Create opacity(unsigned char pOpacity) requires std::derived_from<Node, cocos2d::CCNodeRGBA> {
+        Setup opacity(unsigned char pOpacity) requires std::derived_from<Node, cocos2d::CCRGBAProtocol> {
             m_node->setOpacity(pOpacity);
             return {m_node};
         }
-        Create opacity(cocos2d::CCNodeRGBA* pCopy) requires std::derived_from<Node, cocos2d::CCNodeRGBA> {
+        Setup opacity(Node* pCopy) requires std::derived_from<Node, cocos2d::CCRGBAProtocol> {
             return opacity(pCopy->getOpacity());
         }
 
 
-
-        Create flipX(bool pFlip) requires std::derived_from<Node, cocos2d::CCSprite> {
-            m_node->setFlipX(pFlip);
-            return {m_node};
-        }
-        Create flipY(bool pFlip) requires std::derived_from<Node, cocos2d::CCSprite> {
-            m_node->setFlipY(pFlip);
-            return {m_node};
-        }
-
-
         
-        Create layout(geode::Layout* pLayout) requires std::derived_from<Node, cocos2d::CCMenu> {
+        Setup layout(geode::Layout* pLayout) requires std::derived_from<Node, cocos2d::CCMenu> {
             m_node->setLayout(pLayout);
             return {m_node};
         }
-        Create updateLayout() requires std::derived_from<Node, cocos2d::CCMenu> {
+        Setup updateLayout() requires std::derived_from<Node, cocos2d::CCMenu> {
             m_node->updateLayout();
             return {m_node};
         }
 
 
 
-        Create callback(cocos2d::CCObject* pTarget, cocos2d::SEL_MenuHandler pSelector) requires std::derived_from<Node, cocos2d::CCMenuItem> {
+        Setup callback(cocos2d::CCObject* pTarget, cocos2d::SEL_MenuHandler pSelector) requires std::derived_from<Node, cocos2d::CCMenuItem> {
             m_node->setTarget(pTarget, pSelector);
             return {m_node};
         }
-        Create callback(geode::Function<void(Node*)> pCallback) requires std::derived_from<Node, cocos2d::CCMenuItem> {
+        Setup callback(geode::Function<void(Node*)> pCallback) requires std::derived_from<Node, cocos2d::CCMenuItem> {
             geode::cocos::CCMenuItemExt::assignCallback(m_node, std::move(pCallback));
             return {m_node};
         }
-        Create callback(TextInputCallback pCallback) requires std::derived_from<Node, geode::TextInput> {
+        Setup callback(TextInputCallback pCallback) requires std::derived_from<Node, geode::TextInput> {
             m_node->setCallback(std::move(pCallback));
             return {m_node};
         }
 
 
         
-        Create string(geode::ZStringView pString) requires requires { m_node->setString(""); } {
+        Setup string(geode::ZStringView pString) requires requires { m_node->setString(""); } {
             m_node->setString(pString.c_str());
             return {m_node};
         }
         
         
         
-        Create filter(geode::ZStringView pFilter) requires std::derived_from<Node, geode::TextInput> {
+        Setup filter(geode::ZStringView pFilter) requires std::derived_from<Node, geode::TextInput> {
             m_node->setFilter(pFilter);
             return {m_node};
         }
-        Create filter(geode::CommonFilter pFilter) requires std::derived_from<Node, geode::TextInput> {
+        Setup filter(geode::CommonFilter pFilter) requires std::derived_from<Node, geode::TextInput> {
             m_node->setCommonFilter(pFilter);
             return {m_node};
         }
-        Create maxCharCount(std::size_t pLength) requires std::derived_from<Node, geode::TextInput> {
+        Setup maxCharCount(std::size_t pLength) requires std::derived_from<Node, geode::TextInput> {
             m_node->setMaxCharCount(pLength);
             return {m_node};
         }
-        Create placeholder(geode::ZStringView pPlaceholder) requires std::derived_from<Node, geode::TextInput> {
+        Setup placeholder(geode::ZStringView pPlaceholder) requires std::derived_from<Node, geode::TextInput> {
             m_node->setPlaceholder(pPlaceholder);
+            return {m_node};
+        }
+
+
+
+        Setup toggle(bool pOn) requires std::derived_from<Node, CCMenuItemToggler> {
+            m_node->toggle(pOn);
             return {m_node};
         }
     };
 
     template<typename T>
-    auto node(Create<T> pCreate) {
-        return pCreate.get();
+    Setup(T*) -> Setup<T*, T>;
+
+    template<typename T>
+    auto node(Setup<T> pSetup) {
+        return pSetup.get();
     }
 }